Flight_price/requirements.txt
streamlit
pandas
scikit-learn
python-dotenv
numpy
matplotlib
xgboost
mlflow